src/main/java/baseball/generator/RandomNumGenerator.java
package baseball.generator;

import camp.nextstep.edu.missionutils.Randoms;

import java.util.ArrayList;
import java.util.List;

public class RandomNumGenerator {

    private RandomNumGenerator() {
    }

    public static List<Integer> generate() {
        List<Integer> numbers = new ArrayList<>();
        while (numbers.size() < 3) {
            int randomNumber = drawRandomNumber();
            if (!numbers.contains(randomNumber)) {
                numbers.add(randomNumber);
            }
        }
        return numbers;
    }

    private static int drawRandomNumber() {
        return Randoms.pickNumberInRange(1, 9);
    }
}
